+++ lib/data/model/hour_model.dart
@@ -15,7 +15,7 @@ class HourModel extends Hour {
           tempC: json['temp_c'],
           condition: ConditionModel.fromJson(json['condition']),
           windKph: json['wind_kph'],
-          humidity: json['humidity'],
+          humidity: json['humidity'].toDouble(),
           feelslikeC: json['feelslike_c'],
           uv: json['uv']);
 }
